QVariant: added toFloat and toReal
Made better use of qreal all over the place. We were previously
using QVariant::toDouble a lot. That is triggering unnecessary
conversions between float and double on embedded.

+++ src/gui/image/qpnghandler.cpp
@@ -956,7 +956,7 @@ QVariant QPngHandler::option(ImageOption option) const
 void QPngHandler::setOption(ImageOption option, const QVariant &value)
 {
     if (option == Gamma)
-        d->gamma = value.toDouble();
+        d->gamma = value.toFloat();
     else if (option == Quality)
         d->quality = value.toInt();
     else if (option == Description)
